move some functions in separate files (floyd warshall array)
