+++ floyd_warshall_array.c
@@ -5,24 +5,28 @@
 #include <math.h>
 #include <stdbool.h>
 
-#define INF __INT16_MAX__
+// #include "include/adj_matrix_utils.h"
+# include "src/adj_matrix_utils.cpp"
 
-#define DENSITY 60 //%
-#define MIN_COST 1
-#define MAX_COST 20
+// #define INF __INT16_MAX__
+
+// #define DENSITY 60 //%
+// #define MIN_COST 1
+// #define MAX_COST 20
 
 #define min(a,b) ((a < b) ? a : b)
-#define bool_to_string(cond) (cond ? "true" : "false")
+// #define bool_to_string(cond) (cond ? "true" : "false")
 
-void print_matrix(int *matrix, int m, int n);
-void print_array(int *array, int size);
-void print_element(int val, int infinity);
+// void print_arr_matrix(int *matrix, int m, int n);
+// void print_array(int *array, int size);
+// void print_element(int val, int infinity);
 void floyd_warshall(int *matrix, int n);
 void floyd_warshall_blocked(int *matrix, int n, int B);
 int sum_if_not_infinite(int a, int b, int infinity);
 void execute_round(int *matrix, int n, int t, int row, int col, int B);
-void generate_graph(int *matrix, int n, int seed);
-bool same_matrix(int *matrix_1, int *matrix_2, int n);
+
+// void generate_arr_graph(int *matrix, int n, int seed);
+// bool same_arr_matrix(int *matrix_1, int *matrix_2, int n);
 
 int main() {
 
@@ -43,60 +47,60 @@ int main() {
     srand(rand_seed);
 
     //matrix initialization
-    generate_graph(rand_matrix_1, n, rand_seed);
+    generate_arr_graph(rand_matrix_1, n, rand_seed);
 
     //matrix print
     printf("input adjacency matrix %lux%lu:\n", n, n);
-    print_matrix(rand_matrix_1, n, n);
+    print_arr_matrix(rand_matrix_1, n, n);
 
     //floyd_warshall execution
     floyd_warshall(rand_matrix_1, n);
 
     //print floyd_warshall output
     printf("output adjacency matrix classic %lux%lu:\n", n, n);
-    print_matrix(rand_matrix_1, n, n);
+    print_arr_matrix(rand_matrix_1, n, n);
 
     //---------------------------------------------------------------
 
     //matrix initialization with same seed
-    generate_graph(rand_matrix_2, n, rand_seed);
+    generate_arr_graph(rand_matrix_2, n, rand_seed);
     
     //floyd_warshall_blocked execution
     floyd_warshall_blocked(rand_matrix_2, n, BLOCKING_FACTOR);
     
     //print floyd_warshall_blocked output
     printf("output adjacency matrix blocked %lux%lu:\n", n, n);
-    print_matrix(rand_matrix_2, n, n);
+    print_arr_matrix(rand_matrix_2, n, n);
 
     //---------------------------------------------------------------
 
     //compare matrixes output
-    bool are_the_same = same_matrix(rand_matrix_1, rand_matrix_2, n);
+    bool are_the_same = same_arr_matrix(rand_matrix_1, rand_matrix_2, n);
     printf("Matrixes are equal? %s\n", bool_to_string(are_the_same));
 
     return 0;
 }
 
-bool same_matrix(int *matrix_1, int *matrix_2, int n) {
-    for (int i = 0; i < n; i++) {
-        if(matrix_1[i] != matrix_2[i]) return false;
-    }
-    return true;
-}
+// bool same_arr_matrix(int *matrix_1, int *matrix_2, int n) {
+//     for (int i = 0; i < n; i++) {
+//         if(matrix_1[i] != matrix_2[i]) return false;
+//     }
+//     return true;
+// }
 
-void generate_graph(int *matrix, int n, int seed) {
-    srand(seed);
-    for (int i = 0; i < n; i++) {
-        matrix[i*n + i] = 0;
-        for (int j = i+1; j < n; j++) {
-            bool add_edge = (rand() % 100) <= DENSITY;
-            int val = (rand() % MAX_COST) + MIN_COST;
-            matrix[i*n + j] = add_edge ? val : INF;
-            //non-oriented graph
-            matrix[j*n + i] = matrix[i*n + j];
-        }
-    }
-}
+// void generate_arr_graph(int *matrix, int n, int seed) {
+//     srand(seed);
+//     for (int i = 0; i < n; i++) {
+//         matrix[i*n + i] = 0;
+//         for (int j = i+1; j < n; j++) {
+//             bool add_edge = (rand() % 100) <= DENSITY;
+//             int val = (rand() % MAX_COST) + MIN_COST;
+//             matrix[i*n + j] = add_edge ? val : INF;
+//             //non-oriented graph
+//             matrix[j*n + i] = matrix[i*n + j];
+//         }
+//     }
+// }
 
 void floyd_warshall(int *matrix, int n) {
     for(int k = 0; k < n; k++) {
@@ -187,7 +191,7 @@ void execute_round(int *matrix, int n, int t, int row, int col, int B) {
                 int x2 =  matrix[k*n + j];
                 int b = sum_if_not_infinite(matrix[i*n + k], matrix[k*n + j], INF);
                 matrix[i*n + j] = min(a, b);
-                //print_matrix(matrix, n, n);
+                //print_arr_matrix(matrix, n, n);
             }
         }
     }
@@ -198,27 +202,27 @@ int sum_if_not_infinite(int a, int b, int infinity) {
     return isInf ? infinity : a+b;
 }
 
-void print_matrix(int *matrix, int m, int n) {
-    printf("[\n");
-    for (int i = 0; i < m; i++) {
-        printf("  ");
-        print_array(&(matrix[i]), n);
-    }
-    printf("]\n");
-}
+// void print_arr_matrix(int *matrix, int m, int n) {
+//     printf("[\n");
+//     for (int i = 0; i < m; i++) {
+//         printf("  ");
+//         print_array(&(matrix[i]), n);
+//     }
+//     printf("]\n");
+// }
 
-void print_array(int *array, int size) {
-    printf("[");
-    for (int i = 0; i < size; i++) {
-        print_element(array[i], INF);
-        if (i < size-1) printf(", ");
-    }
-    printf("]\n");
-}
+// void print_array(int *array, int size) {
+//     printf("[");
+//     for (int i = 0; i < size; i++) {
+//         print_element(array[i], INF);
+//         if (i < size-1) printf(", ");
+//     }
+//     printf("]\n");
+// }
 
-void print_element(int val, int infinity) {
-    if (val < infinity)
-        printf("%02d", val);
-    else 
-        printf("--");
-}
+// void print_element(int val, int infinity) {
+//     if (val < infinity)
+//         printf("%02d", val);
+//     else 
+//         printf("--");
+// }
