+++ floyd_warshall_matrix.c
@@ -8,23 +8,12 @@
 // #include "include/adj_matrix_utils.h"
 # include "src/adj_matrix_utils.cpp"
 
-// #define INF __INT16_MAX__
-
-// #define DENSITY 60 //%
-// #define MIN_COST 1
-// #define MAX_COST 20
-
 #define min(a,b) ((a < b) ? a : b)
-// #define bool_to_string(cond) (cond ? "true" : "false")
 
-// void print_array(int *array, int size);
-// void print_matrix(int **matrix, int m, int n);
 void floyd_warshall(int **matrix, int n);
 void floyd_warshall_blocked(int **matrix, int n, int B);
 int sum_if_not_infinite(int a, int b, int infinity);
 void execute_round(int **matrix, int n, int t, int row, int col, int B);
-// void generate_graph(int **matrix, int n, int seed);
-// bool same_matrix(int **matrix_1, int **matrix_2, int m, int n);
 
 int main() {
 
@@ -86,29 +75,6 @@ int main() {
 
     return 0;
 }
-
-// bool same_matrix(int **matrix_1, int **matrix_2, int m, int n) {
-//     for (int i = 0; i < m; i++) {
-//         for (int j = 0; j < n; j++) {
-//             if(matrix_1[i][j] != matrix_2[i][j]) return false;
-//         }
-//     }
-//     return true;
-// }
-
-// void generate_graph(int **matrix, int n, int seed) {
-//     srand(seed);
-//     for (int i = 0; i < n; i++) {
-//         matrix[i][i] = 0;
-//         for (int j = i+1; j < n; j++) {
-//             bool add_edge = (rand() % 100) <= DENSITY;
-//             int val = (rand() % MAX_COST) + MIN_COST;
-//             matrix[i][j] = add_edge ? val : INF;
-//             //non-oriented graph
-//             matrix[j][i] = matrix[i][j];
-//         }
-//     }
-// }
 
 void floyd_warshall(int **matrix, int n) {
     for(int k = 0; k < n; k++) {
@@ -209,25 +175,3 @@ int sum_if_not_infinite(int a, int b, int infinity) {
     bool isInf = (a == infinity) || (b == infinity);
     return isInf ? infinity : a+b;
 }
-
-// void print_matrix(int **matrix, int m, int n) {
-//     printf("[\n");
-//     for (int i = 0; i < m; i++) {
-//         printf(" ");
-//         print_array(matrix[i], n);
-
-//     }
-//     printf("]\n");
-// }
-
-// void print_array(int *array, int size) {
-//     printf("[");
-//     for (int i = 0; i < size; i++) {
-//         if (array[i] < INF)
-//             printf("%02d", array[i]);
-//         else 
-//             printf("--");
-//         if (i < size-1) printf(", ");
-//     }
-//     printf("]\n");
-// }
