+++ include/adj_matrix_utils.h
@@ -18,7 +18,13 @@
 
 void print_array(int *array, int size);
 void print_matrix(int **matrix, int m, int n);
+void print_element(int val, int infinity);
+
 void generate_graph(int **matrix, int n, int seed);
 bool same_matrix(int **matrix_1, int **matrix_2, int m, int n);
+
+void print_arr_matrix(int *matrix, int m, int n);
+void generate_arr_graph(int *matrix, int n, int seed);
+bool same_arr_matrix(int *matrix_1, int *matrix_2, int n);
 
 #endif // ADJ_MATRIX_UTILS_H
