+++ src/adj_matrix_utils.cpp
@@ -3,28 +3,36 @@
 #include <stdio.h>
 #include <stdlib.h>
 
+// ---------------------------------------------------------------
+//  PRINT UTILS
+
+void print_matrix(int **matrix, int m, int n) {
+    printf("[\n");
+    for (int i = 0; i < m; i++) {
+        printf("  ");
+        print_array(matrix[i], n);
+    }
+    printf("]\n");
+}
 
 void print_array(int *array, int size) {
     printf("[");
     for (int i = 0; i < size; i++) {
-        if (array[i] < INF)
-            printf("%02d", array[i]);
-        else 
-            printf("--");
+        print_element(array[i], INF);
         if (i < size-1) printf(", ");
     }
     printf("]\n");
 }
 
-void print_matrix(int **matrix, int m, int n) {
-    printf("[\n");
-    for (int i = 0; i < m; i++) {
-        printf(" ");
-        print_array(matrix[i], n);
-
-    }
-    printf("]\n");
+void print_element(int val, int infinity) {
+    if (val < infinity)
+        printf("%02d", val);
+    else 
+        printf("--");
 }
+
+// ---------------------------------------------------------------
+// MATRIX GENERATION, COMPARE and others utils
 
 bool same_matrix(int **matrix_1, int **matrix_2, int m, int n) {
     for (int i = 0; i < m; i++) {
@@ -45,6 +53,39 @@ void generate_graph(int **matrix, int n, int seed) {
             matrix[i][j] = add_edge ? val : INF;
             //non-oriented graph
             matrix[j][i] = matrix[i][j];
+        }
+    }
+}
+
+// ---------------------------------------------------------------
+// ARRAY MATRIX FUNCTIONS VARIANTS
+
+void print_arr_matrix(int *matrix, int m, int n) {
+    printf("[\n");
+    for (int i = 0; i < m; i++) {
+        printf("  ");
+        print_array(&(matrix[i]), n);
+    }
+    printf("]\n");
+}
+
+bool same_arr_matrix(int *matrix_1, int *matrix_2, int n) {
+    for (int i = 0; i < n; i++) {
+        if(matrix_1[i] != matrix_2[i]) return false;
+    }
+    return true;
+}
+
+void generate_arr_graph(int *matrix, int n, int seed) {
+    srand(seed);
+    for (int i = 0; i < n; i++) {
+        matrix[i*n + i] = 0;
+        for (int j = i+1; j < n; j++) {
+            bool add_edge = (rand() % 100) <= DENSITY;
+            int val = (rand() % MAX_COST) + MIN_COST;
+            matrix[i*n + j] = add_edge ? val : INF;
+            //non-oriented graph
+            matrix[j*n + i] = matrix[i*n + j];
         }
     }
 }
